refactor: use snprintf and add errors.h module
